refactor: mv package of LineLengthLimitWriter to util
